hfsuser: ensure HFSTIMETOEPOCH returns a 64-bit value on all platforms
The max possible unsigned value for this is 2,212,122,495 or just above
the maximum signed 32 bit int. Explicitly casting to int64 prevents
overflow on platforms like Windows where long is 32 bits.

+++ lib/libhfsuser/hfsuser.h
@@ -11,7 +11,7 @@
 
 #include "libhfs.h"
 
-#define HFSTIMETOEPOCH(x) ((x)-2082844800L)
+#define HFSTIMETOEPOCH(x) ((int64_t)(x)-2082844800)
 
 // maximum bytes an HFS+ path element can occupy in UTF-8
 #define HFS_NAME_MAX 765
